bugfix: avoid repeated status queries for "file(s) not in client view"
Unity will request the version control status of files that are not in the
client view every single frame, which results in a very sluggish editor
experience.  This change introduces a local set to track files that
are known not to be in the client view based on a previous status
query, and omits those files from future status queries.  As a result
the Unity editor remains responsive when editing such assets.